Migrate away from minima theme
Copied the minima theme dirs (_includes, _layouts, _sass) and merged
the assets dir into resources. This allows for customization of the
layout (and I was able to disable the rss feed this way too). This
commit also modifies the old pages to fit into the jekyll way of
doing things. Still lots of work to go to get this clean